+++ src/ufuncs.c
@@ -730,7 +730,7 @@ static PyUFuncGenericFunction Y_d_funcs[1] = {&Y_d_func};
 /* Define the geom -> int functions (Y_i) */
 static void *get_type_id_data[1] = {GEOSGeomTypeId_r};
 static void *get_dimensions_data[1] = {GEOSGeom_getDimensions_r};
-static void *get_coordinate_dimensions_data[1] = {GEOSGeom_getCoordinateDimension_r};
+static void *get_coordinate_dimension_data[1] = {GEOSGeom_getCoordinateDimension_r};
 static void *get_srid_data[1] = {GEOSGetSRID_r};
 static int GetNumPoints(void *context, void *geom, int n) {
     char typ = GEOSGeomTypeId_r(context, geom);
@@ -1901,7 +1901,7 @@ int init_ufuncs(PyObject *m, PyObject *d)
 
     DEFINE_Y_i (get_type_id);
     DEFINE_Y_i (get_dimensions);
-    DEFINE_Y_i (get_coordinate_dimensions);
+    DEFINE_Y_i (get_coordinate_dimension);
     DEFINE_Y_i (get_srid);
     DEFINE_Y_i (get_num_points);
     DEFINE_Y_i (get_num_interior_rings);